+++ Source/Events.Store.MongoDB/Streams/EventsToStreamsWriter.cs
@@ -11,6 +11,7 @@ using Dolittle.Runtime.Events.Store.MongoDB.Events;
 using Dolittle.Runtime.Events.Store.Streams;
 using Microsoft.Extensions.Logging;
 using MongoDB.Driver;
+using StreamEvent = Dolittle.Runtime.Events.Store.MongoDB.Events.StreamEvent;
 
 namespace Dolittle.Runtime.Events.Store.MongoDB.Streams;
 
@@ -20,7 +21,6 @@ namespace Dolittle.Runtime.Events.Store.MongoDB.Streams;
 [DisableAutoRegistration]
 public class EventsToStreamsWriter : IWriteEventsToStreamCollection, IWriteEventsToStreams
 {
-    readonly FilterDefinitionBuilder<Events.StreamEvent> _streamFilter = Builders<Events.StreamEvent>.Filter;
     readonly IStreams _streams;
     readonly IEventConverter _eventConverter;
     readonly IStreamEventWatcher _streamWatcher;
@@ -42,98 +42,155 @@ public class EventsToStreamsWriter : IWriteEventsToStreamCollection, IWriteEvent
     }
 
     /// <inheritdoc/>
-    public async Task Write(CommittedEvent @event, ScopeId scope, StreamId stream, PartitionId partition, CancellationToken cancellationToken)
-    {
-        _logger.WritingEventToStream(@event.EventLogSequenceNumber, stream, scope);
-        var writtenStreamPosition = await Write(
-            await _streams.Get(scope, stream, cancellationToken).ConfigureAwait(false),
-            _streamFilter,
-            streamPosition =>
-                @event is CommittedExternalEvent externalEvent ?
-                    _eventConverter.ToStoreStreamEvent(externalEvent, streamPosition, partition)
-                    : _eventConverter.ToStoreStreamEvent(@event, streamPosition, partition),
-            cancellationToken).ConfigureAwait(false);
-        _streamWatcher.NotifyForEvent(scope, stream, writtenStreamPosition);
-    }
+    public Task Write(CommittedEvent @event, ScopeId scope, StreamId stream, PartitionId partition, CancellationToken cancellationToken)
+        => Write(new []{(@event, partition)}, scope, stream, cancellationToken);
 
     /// <inheritdoc />
     public async Task Write(IEnumerable<(CommittedEvent, PartitionId)> events, ScopeId scope, StreamId stream, CancellationToken cancellationToken)
     {
-        var writtenStreamPosition = await Write(
+        var lastWrittenStreamPosition = await Write(
             await _streams.Get(scope, stream, cancellationToken).ConfigureAwait(false),
-            _streamFilter,
             streamPosition =>
             {
-                return events.Select((eventAndPartition, i) =>
-                    eventAndPartition.Item1 is CommittedExternalEvent externalEvent ?
-                        _eventConverter.ToStoreStreamEvent(externalEvent, streamPosition + (ulong)i, eventAndPartition.Item2)
-                        : _eventConverter.ToStoreStreamEvent(eventAndPartition.Item1, streamPosition + (ulong)i, eventAndPartition.Item2));
+                return events
+                    .Select((eventAndPartition, i) =>
+                        eventAndPartition.Item1 is CommittedExternalEvent externalEvent
+                            ? _eventConverter.ToStoreStreamEvent(externalEvent, streamPosition + (ulong)i, eventAndPartition.Item2)
+                            : _eventConverter.ToStoreStreamEvent(eventAndPartition.Item1, streamPosition + (ulong)i, eventAndPartition.Item2))
+                    .ToList();
             },
             cancellationToken).ConfigureAwait(false);
-        _streamWatcher.NotifyForEvent(scope, stream, writtenStreamPosition);
+        _streamWatcher.NotifyForEvent(scope, stream, lastWrittenStreamPosition);
     }
 
     /// <inheritdoc/>
-    public async Task<StreamPosition> Write<TEvent>(
-        IMongoCollection<TEvent> stream,
-        FilterDefinitionBuilder<TEvent> filter,
-        Func<StreamPosition, TEvent> createStoreEvent,
-        CancellationToken cancellationToken)
-        where TEvent : class
-    {
-        StreamPosition streamPosition = null;
-        try
-        {
-            using var session = await _streams.StartSessionAsync().ConfigureAwait(false);
-            await session.WithTransactionAsync(
-                async (transaction, cancellationToken) =>
-                {
-                    streamPosition = (ulong)await stream.CountDocumentsAsync(
-                        transaction,
-                        filter.Empty).ConfigureAwait(false);
-
-                    await stream.InsertOneAsync(
-                        transaction,
-                        createStoreEvent(streamPosition),
-                        cancellationToken: cancellationToken).ConfigureAwait(false);
-                    return Task.CompletedTask;
-                },
-                cancellationToken: cancellationToken).ConfigureAwait(false);
-            return streamPosition;
-        }
-        catch (MongoWaitQueueFullException ex)
-        {
-            throw new EventStoreUnavailable("Mongo wait queue is full", ex);
-        }
-    }
+    public Task<StreamPosition> Write<TEvent>(IMongoCollection<TEvent> stream, Func<StreamPosition, TEvent> createStoreEvent, CancellationToken cancellationToken)
+        => WriteOnlyNewEvents(stream, position => new List<TEvent>{createStoreEvent(position)}, cancellationToken);
 
     /// <inheritdoc />
-    public async Task<StreamPosition> Write<TEvent>(IMongoCollection<TEvent> stream, FilterDefinitionBuilder<TEvent> filter, Func<StreamPosition, IEnumerable<TEvent>> createStoreEvents, CancellationToken cancellationToken)
-        where TEvent : class
+    public Task<StreamPosition> Write<TEvent>(IMongoCollection<TEvent> stream, Func<StreamPosition, IReadOnlyList<TEvent>> createStoreEvents, CancellationToken cancellationToken)
+        => WriteOnlyNewEvents(stream, createStoreEvents, cancellationToken);
+
+    async Task<StreamPosition> WriteOnlyNewEvents<TEvent>(IMongoCollection<TEvent> stream, Func<StreamPosition, IReadOnlyList<TEvent>> createEventsToWrite, CancellationToken cancellationToken)
     {
-        StreamPosition streamPosition = null;
+        Task WriteToCollection(IReadOnlyList<TEvent> events, IClientSessionHandle transaction, CancellationToken cancellationToken) => events.Count switch
+        {
+            0 => Task.CompletedTask,
+            1 => stream.InsertOneAsync(transaction, events[0], cancellationToken: cancellationToken),
+            _ => stream.InsertManyAsync(transaction, events, cancellationToken: cancellationToken)
+        };
+
         try
         {
             using var session = await _streams.StartSessionAsync(cancellationToken: cancellationToken).ConfigureAwait(false);
-            return await session.WithTransactionAsync(
-                async (transaction, cancellationToken) =>
-                {
-                    streamPosition = (ulong)await stream.CountDocumentsAsync(
-                        transaction,
-                        filter.Empty, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                    var eventsToWrite = createStoreEvents(streamPosition).ToArray(); 
-                    await stream.InsertManyAsync(
-                        transaction,
-                        eventsToWrite,
-                        cancellationToken: cancellationToken).ConfigureAwait(false);
-                    return streamPosition + (ulong)(eventsToWrite.Length - 1);
-                },
-                cancellationToken: cancellationToken).ConfigureAwait(false);
+            try
+            {
+                session.StartTransaction();
+                var streamPosition = (ulong) await stream.CountDocumentsAsync(
+                    session,
+                    Builders<TEvent>.Filter.Empty,
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
+                var eventsToWrite = createEventsToWrite(streamPosition).ToList();
+                await WriteToCollection(eventsToWrite, session, cancellationToken).ConfigureAwait(false);
+                await session.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
+                return streamPosition + (ulong) (eventsToWrite.Count - 1);
+            }
+            catch
+            {
+                await session.AbortTransactionAsync(cancellationToken).ConfigureAwait(false);
+                throw;
+            }
         }
         catch (MongoWaitQueueFullException ex)
         {
             throw new EventStoreUnavailable("Mongo wait queue is full", ex);
         }
+        catch (Exception ex) when (typeof(TEvent) == typeof(Events.StreamEvent) && IsDuplicateKeyException(ex))
+        {
+            using var session = await _streams.StartSessionAsync(cancellationToken: cancellationToken).ConfigureAwait(false);
+            try
+            {
+                session.StartTransaction();
+                var streamPosition = (ulong) await stream.CountDocumentsAsync(
+                    session,
+                    Builders<TEvent>.Filter.Empty,
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
+                var eventsToWrite = createEventsToWrite(streamPosition).ToList();
+                var streamEventsToWrite = eventsToWrite.Cast<Events.StreamEvent>().ToList();
+                var storedEvents = await GetStoredEventStreamHeadOfStreamToWrite((IMongoCollection<Events.StreamEvent>) stream, streamEventsToWrite, session, cancellationToken).ConfigureAwait(false);
+        
+                if (!CheckIfStoredEventsIsPrefixOfEventsToWrite(streamEventsToWrite, storedEvents))
+                {
+                    throw;
+                }
+        
+                eventsToWrite = eventsToWrite.Skip(storedEvents.Count).ToList();
+                await WriteToCollection(eventsToWrite, session, cancellationToken).ConfigureAwait(false);
+        
+                await session.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
+                return streamPosition + (ulong) (eventsToWrite.Count - 1);
+            }
+            catch
+            {
+                await session.AbortTransactionAsync(cancellationToken).ConfigureAwait(false);
+                throw;
+            }
+        }
     }
+
+    static Task<List<Events.StreamEvent>> GetStoredEventStreamHeadOfStreamToWrite(IMongoCollection<Events.StreamEvent> stream, IReadOnlyList<Events.StreamEvent> eventsToWrite, IClientSessionHandle transaction, CancellationToken cancellationToken)
+        => stream
+            .Find(
+                transaction,
+                Builders<Events.StreamEvent>.Filter.Gte(_ => _.Metadata.EventLogSequenceNumber, eventsToWrite[0].Metadata.EventLogSequenceNumber))
+            .SortBy(_ => _.Metadata.EventLogSequenceNumber)
+            .ToListAsync(cancellationToken);
+
+    static bool CheckIfStoredEventsIsPrefixOfEventsToWrite(IReadOnlyList<Events.StreamEvent> eventsToWrite, IReadOnlyList<Events.StreamEvent> storedEvents)
+    {
+        if (storedEvents.Count == 0)
+        {
+            return true;
+        }
+
+        if (storedEvents.Count > eventsToWrite.Count)
+        {
+            return false;
+        }
+        
+        ThrowIfStoredEventsIsUnOrdered(storedEvents);
+        for (var i = 0; i < storedEvents.Count; i++)
+        {
+            if (!IsSameStreamEvents(eventsToWrite[i], storedEvents[i]))
+            {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    static void ThrowIfStoredEventsIsUnOrdered(IReadOnlyList<StreamEvent> storedEvents)
+    {
+        for (var i = 0; i < storedEvents.Count - 1; i++)
+        {
+            if (storedEvents[i + 1].Metadata.EventLogSequenceNumber < storedEvents[i].Metadata.EventLogSequenceNumber)
+            {
+                throw new StoredStreamEventsHeadIsNotOrdered();
+            }
+        }
+    }
+
+    static bool IsSameStreamEvents(Events.StreamEvent eventToWrite, Events.StreamEvent storedEvent)
+        => eventToWrite.Metadata.EventLogSequenceNumber.Equals(storedEvent.Metadata.EventLogSequenceNumber)
+            && eventToWrite.Partition.Equals(storedEvent.Partition);
+
+    static bool IsDuplicateKeyException(Exception exception)
+        => exception switch
+        {
+            MongoDuplicateKeyException => true,
+            MongoWriteException writeException => writeException.Message.Contains("duplicate key error"),
+            MongoBulkWriteException bulkWriteException => bulkWriteException.Message.Contains("duplicate key error"),
+            _ => false,
+        };
 }